requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
cryptography>=43.0.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
GitPython==3.1.41
python-dotenv==1.0.1
mkdocs>=1.5.0
mkdocs-material>=9.4.0
pytest>=8.0.0
openai>=1.0.0
psycopg2-binary==2.9.9
PyJWT>=2.8.0
google-auth>=2.29.0
